requirements.txt
opencv-python
pypylon
matplotlib
Pillow